# .schema/api.swagger.json
{
  "consumes": [
    "application/json",
    "application/x-www-form-urlencoded"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Welcome to the ORY Kratos HTTP API documentation!",
    "title": "Ory Kratos",
    "version": "latest"
  },
  "basePath": "/",
  "paths": {
...
    "/self-service/registration/api": {
      "get": {
        "description": "This endpoint initiates a registration flow for API clients such as mobile devices, smart TVs, and so on.\n\nIf a valid provided session cookie or session token is provided, a 400 Bad Request error\nwill be returned unless the URL query parameter `?refresh=true` is set.\n\nTo fetch an existing registration flow call `/self-service/registration/flows?flow=\u003cflow_id\u003e`.\n\n:::warning\n\nYou MUST NOT use this endpoint in client-side (Single Page Apps, ReactJS, AngularJS) nor server-side (Java Server\nPages, NodeJS, PHP, Golang, ...) browser applications. Using this endpoint in these applications will make\nyou vulnerable to a variety of CSRF attacks.\n\nThis endpoint MUST ONLY be used in scenarios such as native mobile apps (React Native, Objective C, Swift, Java, ...).\n\n:::\n\nMore information can be found at [ORY Kratos User Login and User Registration Documentation](https://www.ory.sh/docs/next/kratos/self-service/flows/user-login-user-registration).",
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "public"
        ],
        "summary": "Initialize Registration Flow for API clients",
        "operationId": "initializeSelfServiceRegistrationViaAPIFlow",
        "responses": {
          "200": {
            "description": "registrationFlow",
            "schema": {
              "$ref": "#/definitions/registrationFlow"
            }
          },
          "400": {
            "description": "genericError",
            "schema": {
              "$ref": "#/definitions/genericError"
            }
          },
          "500": {
            "description": "genericError",
            "schema": {
              "$ref": "#/definitions/genericError"
            }
          }
        }
      }
    },
    "/self-service/registration/browser": {
      "get": {
        "description": "This endpoint initializes a browser-based user registration flow. Once initialized, the browser will be redirected to\n`selfservice.flows.registration.ui_url` with the flow ID set as the query parameter `?flow=`. If a valid user session\nexists already, the browser will be redirected to `urls.default_redirect_url` unless the query parameter\n`?refresh=true` was set.\n\n:::note\n\nThis endpoint is NOT INTENDED for API clients and only works with browsers (Chrome, Firefox, ...).\n\n:::\n\nMore information can be found at [ORY Kratos User Login and User Registration Documentation](https://www.ory.sh/docs/next/kratos/self-service/flows/user-login-user-registration).",
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "public"
        ],
        "summary": "Initialize Registration Flow for browsers",
        "operationId": "initializeSelfServiceRegistrationViaBrowserFlow",
        "responses": {
          "302": {
            "description": "Empty responses are sent when, for example, resources are deleted. The HTTP status code for empty responses is\ntypically 201."
          },
          "500": {
            "description": "genericError",
            "schema": {
              "$ref": "#/definitions/genericError"
            }
          }
        }
      }
    },
    "/self-service/registration/flows": {
      "get": {
        "description": "This endpoint returns a registration flow's context with, for example, error details and other information.\n\nMore information can be found at [ORY Kratos User Login and User Registration Documentation](https://www.ory.sh/docs/next/kratos/self-service/flows/user-login-user-registration).",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "common",
          "public",
          "admin"
        ],
        "summary": "Get information about a registration flow",
        "operationId": "getSelfServiceRegistrationFlow",
        "parameters": [
          {
            "type": "string",
            "description": "The Registration Flow ID\n\nThe value for this parameter comes from `flow` URL Query parameter sent to your\napplication (e.g. `/registration?flow=abcde`).",
            "name": "id",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "registrationFlow",
            "schema": {
              "$ref": "#/definitions/registrationFlow"
            }
          },
          "403": {
            "description": "genericError",
            "schema": {
              "$ref": "#/definitions/genericError"
            }
          },
          "404": {
            "description": "genericError",
            "schema": {
              "$ref": "#/definitions/genericError"
            }
          },
          "410": {
            "description": "genericError",
            "schema": {
              "$ref": "#/definitions/genericError"
            }
          },
          "500": {
            "description": "genericError",
            "schema": {
              "$ref": "#/definitions/genericError"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "CredentialsType": {
      "description": "and so on.",
      "type": "string",
      "title": "CredentialsType  represents several different credential types, like password credentials, passwordless credentials,"
    },
...
    "ID": {
      "type": "integer",
      "format": "int64"
    },
...
    "Message": {
      "type": "object",
      "properties": {
        "context": {
          "type": "object"
        },
        "id": {
          "$ref": "#/definitions/ID"
        },
        "text": {
          "type": "string"
        },
        "type": {
          "$ref": "#/definitions/Type"
        }
      }
    },
    "Messages": {
      "type": "array",
      "items": {
        "$ref": "#/definitions/Message"
      }
    },
...
    "Type": {
      "description": "The flow type can either be `api` or `browser`.",
      "type": "string",
      "title": "Type is the flow type."
    },
    "UUID": {
      "type": "string",
      "format": "uuid4"
    },
...
    "formField": {
      "description": "Field represents a HTML Form Field",
      "type": "object",
      "required": [
        "name",
        "type"
      ],
      "properties": {
        "disabled": {
          "description": "Disabled is the equivalent of `\u003cinput {{if .Disabled}}disabled{{end}}\"\u003e`",
          "type": "boolean"
        },
        "messages": {
          "$ref": "#/definitions/Messages"
        },
        "name": {
          "description": "Name is the equivalent of `\u003cinput name=\"{{.Name}}\"\u003e`",
          "type": "string"
        },
        "pattern": {
          "description": "Pattern is the equivalent of `\u003cinput pattern=\"{{.Pattern}}\"\u003e`",
          "type": "string"
        },
        "required": {
          "description": "Required is the equivalent of `\u003cinput required=\"{{.Required}}\"\u003e`",
          "type": "boolean"
        },
        "type": {
          "description": "Type is the equivalent of `\u003cinput type=\"{{.Type}}\"\u003e`",
          "type": "string"
        },
        "value": {
          "description": "Value is the equivalent of `\u003cinput value=\"{{.Value}}\"\u003e`",
          "type": "object"
        }
      }
    },
    "formFields": {
      "description": "Fields contains multiple fields",
      "type": "array",
      "items": {
        "$ref": "#/definitions/formField"
      }
    },
    "genericError": {
      "description": "Error responses are sent when an error (e.g. unauthorized, bad request, ...) occurred.",
      "type": "object",
      "title": "Error response",
      "properties": {
        "error": {
          "$ref": "#/definitions/genericErrorPayload"
        }
      }
    },
    "genericErrorPayload": {
      "type": "object",
      "properties": {
        "code": {
          "description": "Code represents the error status code (404, 403, 401, ...).",
          "type": "integer",
          "format": "int64",
          "example": 404
        },
        "debug": {
          "description": "Debug contains debug information. This is usually not available and has to be enabled.",
          "type": "string",
          "example": "The database adapter was unable to find the element"
        },
        "details": {
          "type": "object",
          "additionalProperties": true
        },
        "message": {
          "type": "string"
        },
        "reason": {
          "type": "string"
        },
        "request": {
          "type": "string"
        },
        "status": {
          "type": "string"
        }
      }
    },
...
    "registrationFlow": {
      "type": "object",
      "required": [
        "id",
        "expires_at",
        "issued_at",
        "request_url",
        "methods"
      ],
      "properties": {
        "active": {
          "$ref": "#/definitions/CredentialsType"
        },
        "expires_at": {
          "description": "ExpiresAt is the time (UTC) when the flow expires. If the user still wishes to log in,\na new flow has to be initiated.",
          "type": "string",
          "format": "date-time"
        },
        "id": {
          "$ref": "#/definitions/UUID"
        },
        "issued_at": {
          "description": "IssuedAt is the time (UTC) when the flow occurred.",
          "type": "string",
          "format": "date-time"
        },
        "messages": {
          "$ref": "#/definitions/Messages"
        },
        "methods": {
          "description": "Methods contains context for all enabled registration methods. If a registration flow has been\nprocessed, but for example the password is incorrect, this will contain error messages.",
          "type": "object",
          "additionalProperties": {
            "$ref": "#/definitions/registrationFlowMethod"
          }
        },
        "request_url": {
          "description": "RequestURL is the initial URL that was requested from ORY Kratos. It can be used\nto forward information contained in the URL's path or query for example.",
          "type": "string"
        },
        "type": {
          "$ref": "#/definitions/Type"
        }
      }
    },
    "registrationFlowMethod": {
      "type": "object",
      "properties": {
        "config": {
          "$ref": "#/definitions/registrationFlowMethodConfig"
        },
        "method": {
          "$ref": "#/definitions/CredentialsType"
        }
      }
    },
    "registrationFlowMethodConfig": {
      "type": "object",
      "required": [
        "action",
        "method",
        "fields"
      ],
      "properties": {
        "action": {
          "description": "Action should be used as the form action URL `\u003cform action=\"{{ .Action }}\" method=\"post\"\u003e`.",
          "type": "string"
        },
        "fields": {
          "$ref": "#/definitions/formFields"
        },
        "messages": {
          "$ref": "#/definitions/Messages"
        },
        "method": {
          "description": "Method is the form method (e.g. POST)",
          "type": "string"
        },
        "providers": {
          "description": "Providers is set for the \"oidc\" registration method.",
          "type": "array",
          "items": {
            "$ref": "#/definitions/formField"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "sessionToken": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
...
}